pyproject.toml
[project]
name = "servernoe-prog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
